jaylib_wrapvec2: don't truncate values
Raylib's Vector2 contains two floats, but Jaylib truncates those to integers
when it returns them (note that they are parsed as floats correctly).

This can lead to some pretty confusing behavior -- not necessarily with any of
the currently implemented functions that return vectors, but this prohibits
wrapping some other Raylib functions that do not always return integers.